refactor: extract Provider.Incidents
Move submit/list/get incident-report functions and their two query
helpers onto Provider.Incidents; facade delegates.
